+++ src/main/java/gg/packetloss/grindstone/highscore/ScoreEntry.java
@@ -1,3 +1,9 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
 package gg.packetloss.grindstone.highscore;
 
 import org.bukkit.Bukkit;
@@ -7,9 +13,9 @@ import java.util.UUID;
 
 public class ScoreEntry {
     private UUID playerID;
-    private int score;
+    private long score;
 
-    public ScoreEntry(UUID playerID, int score) {
+    public ScoreEntry(UUID playerID, long score) {
         this.playerID = playerID;
         this.score = score;
     }
@@ -22,7 +28,7 @@ public class ScoreEntry {
         return Bukkit.getOfflinePlayer(playerID);
     }
 
-    public int getScore() {
+    public long getScore() {
         return score;
     }
 }
